LayoutTests/fast/mediastream/peerconnection-mediastreamlist.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="stylesheet" href="../js/resources/js-test-style.css">
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
</body>
<script>
description("This test confirms that out-of-bounds access of MediaStreamList returns undefined.");

var pc = new webkitPeerConnection00("", function() {});

shouldBeUndefined('pc.localStreams[0xbadc0ded]');

window.successfullyParsed = true;

</script>
<script src="../js/resources/js-test-post.js"></script>
</html>
