LayoutTests/fast/mediastream/peerconnection-mediastreamlist-expected.txt

This test confirms that out-of-bounds access of MediaStreamList returns undefined.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pc.localStreams[0xbadc0ded] is undefined.
PASS successfullyParsed is true

TEST COMPLETE
